Saving progress on dynamic way to define functions for different gamemode types
