+++ sudokuru/app/Components/SudokuBoard/Core/Functions/CellFunctions.ts
@@ -4,10 +4,7 @@ import {
   CellLocation,
 } from "../../../../Functions/LocalDatabase";
 import { HintObjectProps } from "../../SudokuBoard";
-import {
-  doesCellHaveConflict,
-  isBoardDisabled,
-} from "../../SudokuBoardFunctions";
+import { isBoardDisabled } from "../../SudokuBoardFunctions";
 import { isValueCorrect } from "./BoardFunctions";
 
 /**
@@ -241,3 +238,23 @@ export function areCellsInSameColumn(
 ) {
   return currentCellCoordinate.c === selectedCellCoordinate.c;
 }
+
+/**
+ * Checks if a given cell in the puzzle has a conflict with the solution.
+ *
+ * @param r - The row index of the cell.
+ * @param c - The column index of the cell.
+ * @param cell - The cell object containing its type and entry.
+ * @returns True if the cell's entry is incorrect; false otherwise.
+ */
+export const doesCellHaveConflict = (
+  sudokuBoard: SudokuObjectProps,
+  r: number,
+  c: number,
+): boolean => {
+  const cell = sudokuBoard.puzzle[r][c];
+  if (cell.type === "note" || cell.entry === 0) {
+    return false;
+  }
+  return !(sudokuBoard.puzzle[r][c].entry === sudokuBoard.puzzleSolution[r][c]);
+};
